jira-service/requirements.txt
fastapi
uvicorn
celery
redis
psycopg2-binary
httpx
cryptography
pydantic
python-dotenv
